pyproject.toml
[project]
name = "entwine"
version = "0.1.0"
description = "LLM-powered enterprise digital twin simulation"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "litellm>=1.60",
    "qdrant-client>=1.13",
    "openai>=1.60",
    "pydantic>=2.10",
    "pydantic-settings>=2.7",
    "pyyaml>=6.0",
    "typer>=0.15",
    "sse-starlette>=2.2",
    "structlog>=25.1",
    "httpx>=0.28",
]

[project.scripts]
entwine = "entwine.cli.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/entwine"]

[dependency-groups]
dev = [
    "pytest>=8.3",
    "pytest-cov>=6.0",
    "mypy>=1.14",
    "pre-commit>=4.1",
    "types-PyYAML>=6.0",
    "ruff>=0.15.5",
    "pytest-asyncio>=1.3.0",
]

[tool.mypy]
python_version = "3.12"
strict = true
warn_unused_configs = true

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "B",   # flake8-bugbear
    "I",   # isort
    "UP",  # pyupgrade
    "C4",  # flake8-comprehensions
    "SIM", # flake8-simplify
    "PTH", # flake8-use-pathlib
    "RUF", # ruff-specific
]
ignore = ["E501", "B008"]  # B008: typer requires Option() in defaults

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=src/entwine --cov-report=term-missing"
markers = ["integration: marks tests that require external services"]
asyncio_mode = "auto"

[tool.coverage.run]
source = ["src/entwine"]
omit = ["tests/*"]
